requirements.txt
aiosqlite
pillow
python-dotenv
pytesseract
telethon
